% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fenn.R
\name{fenn.default}
\alias{fenn.default}
\title{This is the main function for fitting \code{fenn}.}
\usage{
fenn.default(x, ...)
}
\arguments{
\item{x}{An object of class \code{"fenn"}.}
}
\description{
This is the main function for fitting \code{fenn}.
}
\details{
This function is a method for the generic function
\code{print()} for class \code{"fenn"}.
It can be invoked by calling \code{print(x)} for an
object \code{x} of the appropriate class, or directly by
calling \code{print.fenn(x)} regardless of the
class of the object.
}
