% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda.R
\name{plot.slda}
\alias{plot.slda}
\title{Plot Method for Class \code{slda}.}
\usage{
plot.slda(x, panel = panel.slda, \dots, cex = 0.7, dimen,
                 abbrev = FALSE, xlab = "SLD1", ylab = "SLD2")
}
\arguments{
\item{x}{An object of class \code{"slda"}.}

\item{panel}{The panel function used to plot the data.}

\item{\dots}{Additional arguments to \code{pairs}, \code{ldahist} or \code{eqscplot}.}

\item{cex}{Graphics parameter \code{cex} for labels on plots.}

\item{abbrev}{Whether the group labels are abbreviated on the plots. If \code{abbrev > 0}
this gives \code{minlength} in the call to \code{abbreviate}.}

\item{xlab}{Label for the \code{x} axis.}

\item{ylab}{Label for the \code{y} axis.}
}
\description{
Plots a set of data on one, two or more scaled linear discriminants.
}
\details{
This function is a method for the generic function
\code{plot()} for class \code{"slda"}.
It can be invoked by calling \code{plot(x)} for an
object \code{x} of the appropriate class, or directly by
calling \code{plot.slda(x)} regardless of the
class of the object.

The behaviour is determined by the value of \code{dimen}. For
\code{dimen > 2}, a \code{pairs} plot is used. For \code{dimen = 2}, an
equiscaled scatter plot is drawn. For \code{dimen = 1}, a set of
histograms or density plots are drawn.  Use argument \code{type} to
match \code{"histogram"} or \code{"density"} or \code{"both"}.
}
