% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fenn.R
\name{predict.fenn}
\alias{predict.fenn}
\title{Classify Multivariate Observations}
\usage{
predict.fenn(object, newdata, prior = object$prior, dimen,
                              method = c("knn", "lda"), \dots)
}
\arguments{
\item{object}{Object  of class \code{"fenn"}.}

\item{newdata}{Data frame of cases to be classified or, if \code{object}
has a formula, a data frame with columns of the same names as the
variables used.  A vector will be interpreted
as a row vector.  If newdata is missing, an attempt will be
made to retrieve the data used to fit the \code{fenn} object.}

\item{method}{This determines how the parameter estimation is handled. With \code{"plug-in"}
(the default) the usual unbiased parameter estimates are used and
assumed to be correct. With \code{"debiased"} an unbiased estimator of
the log posterior probabilities is used, and with \code{"predictive"} the
parameter estimates are integrated out using a vague prior.}

\item{\dots}{Arguments based from or to other methods}

\item{reduce.dm}{The dimension of the data will be reduced before performing knn.}
}
\value{
A list with components:
\item{class}{The MAP classification (a factor).}
\item{posterior}{If classification method is lda, the posterior probabilities for the 
                 classes in the tilde transformed space will be returned.}
}
\description{
Classify multivariate observations in conjunction with \code{fenn}, and also
project data onto the scaled linear discriminants.
}
\details{
This function is a method for the generic function \code{predict()} for
class \code{"fenn"}.  It can be invoked by calling \code{predict(x)} for
an object \code{x} of the appropriate class, or directly by calling
\code{predict.fenn(x)} regardless of the class of the object.

Missing values in \code{newdata} are handled by returning \code{NA} if the
scaled linear discriminants cannot be evaluated. If \code{newdata} is omitted and
the \code{na.action} of the fit omitted cases, these will be omitted on the
prediction.

This version centres the scaled linear discriminants so that the
weighted mean (weighted by \code{prior}) of the group centroids is at
the origin.
}
\examples{
tr <- sample(1:50, 25)
train <- rbind(iris3[tr,,1], iris3[tr,,2], iris3[tr,,3])
test <- rbind(iris3[-tr,,1], iris3[-tr,,2], iris3[-tr,,3])
cl <- factor(c(rep("s",25), rep("c",25), rep("v",25)))
z <- fenn(train, cl)
predict(z, test)$class

}
