% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fenn.R
\name{print.fenn}
\alias{print.fenn}
\title{Print Method for Class \code{fenn}.}
\usage{
print.fenn(x, ...)
}
\arguments{
\item{x}{An object of class \code{"fenn"}.}
}
\description{
Prints summary table of fenn fitted object.
}
\details{
This function is a method for the generic function
\code{print()} for class \code{"fenn"}.
It can be invoked by calling \code{print(x)} for an
object \code{x} of the appropriate class, or directly by
calling \code{print.fenn(x)} regardless of the
class of the object.
}
