% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda.R
\name{print.slda}
\alias{print.slda}
\title{Print Method for Class \code{slda}.}
\usage{
print.slda(x, ...)
}
\arguments{
\item{x}{An object of class \code{"slda"}.}
}
\description{
Prints summary table of slda fitted object.
}
\details{
This function is a method for the generic function
\code{print()} for class \code{"slda"}.
It can be invoked by calling \code{print(x)} for an
object \code{x} of the appropriate class, or directly by
calling \code{print.slda(x)} regardless of the
class of the object.
}
