\name{slda-package}
\alias{slda-package}
\docType{package}
\title{Scaled Linear Discriminant Analysis}
\description{This package implements the Scaled Linear Discriminant .Analysis (SLDA) method.
             SLDA is a Von Neumann entropy penalized distance metric learning model
             to scale and project multi-class data on directions of maximal dissimilarity for optimal
             separability. 
}
\details{
\tabular{ll}{
Package: \tab slda\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2018-08-18\cr
License: \tab GPL (>=3)\cr
LazyLoad: \tab yes\cr
}
}
\author{Kourosh Zarringhalam [aut,cre], Carl Fakhry [aut,cre], Sho Inaba [aut,cre]

Maintainers: Kourosh Zarringhalam <kourosh.zarringhalam@umb.edu>, Carl Tony Fakhry <cfakhry@cs.umb.edu>
}
%\references{}