% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda.R
\name{slda.default}
\alias{slda.default}
\title{This is the main function for fitting \code{slda}.}
\usage{
slda.default(x, ...)
}
\arguments{
\item{x}{An object of class \code{"slda"}.}
}
\description{
This is the main function for fitting \code{slda}.
}
\details{
This function is a method for the generic function
\code{print()} for class \code{"slda"}.
It can be invoked by calling \code{print(x)} for an
object \code{x} of the appropriate class, or directly by
calling \code{print.slda(x)} regardless of the
class of the object.
}
